[omm] Move around some util files
